+++ conf/settings.py
@@ -68,6 +68,7 @@ MIDDLEWARE_CLASSES = [
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.middleware.common.CommonMiddleware',
     'directory_components.middleware.NoCacheMiddlware',
+    'directory_components.middleware.CheckGATags'
 ]
 
 ROOT_URLCONF = 'conf.urls'
@@ -99,11 +100,17 @@ TEMPLATES = [
 
 WSGI_APPLICATION = 'conf.wsgi.application'
 
+VCAP_SERVICES = env.json('VCAP_SERVICES', {})
 
-if env.str('REDIS_URL', ''):
+if 'redis' in VCAP_SERVICES:
+    REDIS_URL = VCAP_SERVICES['redis'][0]['credentials']['uri']
+else:
+    REDIS_URL = env.str('REDIS_URL', '')
+
+if REDIS_URL:
     cache = {
         'BACKEND': 'django_redis.cache.RedisCache',
-        'LOCATION': env.str('REDIS_URL'),
+        'LOCATION': REDIS_URL,
         'OPTIONS': {
             'CLIENT_CLASS': "django_redis.client.DefaultClient",
         }
@@ -113,6 +120,7 @@ else:
         'BACKEND': 'django.core.cache.backends.locmem.LocMemCache',
         'LOCATION': 'unique-snowflake',
     }
+
 
 CACHES = {
     'default': cache,
@@ -393,7 +401,16 @@ FEATURE_FLAGS = {
     'MAINTENANCE_MODE_ON': env.bool('FEATURE_MAINTENANCE_MODE_ENABLED', False),
     'RECOMMENDED_FOR_CHOSEN_COUNTRY_ON':
         env.bool('FEATURE_RECOMMENDED_FOR_CHOSEN_COUNTRY_ENABLED', False),
-    'COUNTRY_SELECTOR_ON': env.bool('FEATURE_COUNTRY_SELECTOR_ENABLED', False)
+    'COUNTRY_SELECTOR_ON': env.bool('FEATURE_COUNTRY_SELECTOR_ENABLED', False),
+    'INVESTMENT_SUPPORT_DIRECTORY_LINK_ON': env.bool(
+        'FEATURE_INVESTMENT_SUPPORT_DIRECTORY_LINK_ENABLED', False
+    ),
+    'CAPITAL_INVEST_REGION_SECTOR_OPP_PAGES_ON': env.bool(
+        'FEATURE_CAPITAL_INVEST_REGION_SECTOR_OPP_PAGE_ENABLED', False
+    ),
+    'CAPITAL_INVEST_LANDING_PAGE_ON': env.bool(
+        'FEATURE_CAPITAL_INVEST_LANDING_PAGE_ENABLED', False
+    ),
 }
 
 # Invest High Potential Opportunities