+++ core/views.py
@@ -1,4 +1,5 @@
 from django.conf import settings
+from django.http import Http404
 from django.views.generic import TemplateView
 from django.utils.functional import cached_property
 from django.utils import translation
@@ -18,21 +19,31 @@ from core.context_modifiers import (
     register_context_modifier,
     registry as context_modifier_registry
 )
+from core.helpers import get_ga_data_for_page
 from core.mixins import (
     TEMPLATE_MAPPING, NotFoundOnDisabledFeature, RegionalContentMixin)
 
 
 class CMSPageFromPathView(
-    GA360Mixin,
     RegionalContentMixin,
     CMSLanguageSwitcherMixin,
     NotFoundOnDisabledFeature,
+    GA360Mixin,
     TemplateView
 ):
 
-    @property
-    def ga360_payload(self):
-        return {'page_type': self.page['page_type']}
+    def dispatch(self, request, *args, **kwargs):
+        dispatch_result = super().dispatch(request, *args, **kwargs)
+
+        page_type = self.page['page_type']
+        ga360_data = get_ga_data_for_page(page_type)
+        self.set_ga360_payload(
+            page_id=page_type,
+            business_unit=ga360_data['business_unit'],
+            site_section=ga360_data['site_section'],
+            site_subsection=ga360_data['site_subsection']
+        )
+        return dispatch_result
 
     @property
     def template_name(self):
@@ -52,12 +63,41 @@ class CMSPageFromPathView(
 
         context = super().get_context_data(page=self.page, **kwargs)
 
+        flag_map = {
+            'CapitalInvestRegionPage':
+                'CAPITAL_INVEST_REGION_SECTOR_OPP_PAGES_ON',
+            'CapitalInvestRegionalSectorPage':
+                'CAPITAL_INVEST_REGION_SECTOR_OPP_PAGES_ON',
+            'CapitalInvestOpportunityPage':
+                'CAPITAL_INVEST_REGION_SECTOR_OPP_PAGES_ON',
+            'InternationalCapitalInvestLandingPage':
+                'CAPITAL_INVEST_LANDING_PAGE_ON',
+        }
+
+        flag_name = flag_map.get(self.page['page_type'])
+
+        if flag_name and not settings.FEATURE_FLAGS[flag_name]:
+            raise Http404
+
         for modifier in context_modifier_registry.get_for_page_type(
             self.page['page_type']
         ):
             context.update(modifier(context, request=self.request))
 
         return context
+
+
+class HowToSetupInTheUKView(CMSPageFromPathView):
+    def get_context_data(self, **kwargs):
+        language_code = translation.get_language()
+        show_find_uk_specialist = (
+            settings.FEATURE_FLAGS['INVESTMENT_SUPPORT_DIRECTORY_LINK_ON'] and
+            language_code == settings.LANGUAGE_CODE
+        )
+        return super().get_context_data(
+            show_find_uk_specialist=show_find_uk_specialist,
+            **kwargs
+        )
 
 
 @register_context_modifier('InternationalArticlePage')
@@ -125,8 +165,19 @@ def sector_page_context_modifier(context, request):
         }
 
 
-class InternationalContactPageView(CountryDisplayMixin, TemplateView):
+class InternationalContactPageView(CountryDisplayMixin,
+                                   GA360Mixin,
+                                   TemplateView):
     template_name = 'core/contact_page.html'
+
+    def __init__(self):
+        super().__init__()
+        self.set_ga360_payload(
+            page_id='InternationalContactPage',
+            business_unit='International',
+            site_section='Contact',
+            site_subsection='ContactForm'
+        )
 
     def get_context_data(self, *args, **kwargs):
         return super().get_context_data(
@@ -134,3 +185,40 @@ class InternationalContactPageView(CountryDisplayMixin, TemplateView):
             invest_contact_us_url=urls.build_invest_url('contact/'),
             *args, **kwargs
         )
+
+
+@register_context_modifier('CapitalInvestRegionPage')
+def capital_invest_region_page_context_modifier(context, request):
+
+    def count_data_with_field(list_of_data, field):
+        filtered_list = [item for item in list_of_data if item[field]]
+        return len(filtered_list)
+
+    page = context['page']
+
+    return {
+        'num_of_economics_statistics': count_data_with_field(
+            page['economics_stats'], 'number'),
+        'num_of_location_statistics': count_data_with_field(
+            page['location_stats'], 'number'),
+        'invest_cta_link': urls.SERVICES_INVEST,
+        'buy_cta_link': urls.SERVICES_FAS,
+    }
+
+
+@register_context_modifier('CapitalInvestRegionalSectorPage')
+def capital_invest_regional_sector_page_context_modifier(context, request):
+
+    return {
+        'invest_cta_link': urls.SERVICES_INVEST,
+        'buy_cta_link': urls.SERVICES_FAS,
+    }
+
+
+@register_context_modifier('CapitalInvestOpportunityPage')
+def capital_invest_opportunity_page_context_modifier(context, request):
+
+    return {
+        'invest_cta_link': urls.SERVICES_INVEST,
+        'buy_cta_link': urls.SERVICES_FAS,
+    }